test: merge duplicated pts integration tests suits
